Refactor common actions and types from  sourcecast and sourcereel
