+++ src/features/sourceRecorder/SourceRecorderTypes.ts
@@ -1,16 +1,13 @@
 import { ExternalLibraryName } from '../../commons/application/types/ExternalTypes';
 import { Position } from '../../commons/editor/EditorTypes';
-import { WorkspaceState } from '../../commons/workspace/WorkspaceTypes';
 
-export const FETCH_SOURCECAST_INDEX = 'FETCH_SOURCECAST_INDEX';
+export const SAVE_SOURCECAST_DATA = 'SAVE_SOURCECAST_DATA';
 export const SET_CURRENT_PLAYER_TIME = 'SET_CURRENT_PLAYER_TIME';
 export const SET_CODE_DELTAS_TO_APPLY = 'SET_CODE_DELTAS_TO_APPLY';
-export const SET_EDITOR_READONLY = 'SET_EDITOR_READONLY';
 export const SET_INPUT_TO_APPLY = 'SET_INPUT_TO_APPLY';
 export const SET_SOURCECAST_DATA = 'SET_SOURCECAST_DATA';
 export const SET_SOURCECAST_PLAYBACK_DURATION = 'SET_SOURCECAST_PLAYBACK_DURATION';
 export const SET_SOURCECAST_PLAYBACK_STATUS = 'SET_SOURCECAST_PLAYBACK_STATUS';
-export const UPDATE_SOURCECAST_INDEX = 'UPDATE_SOURCECAST_INDEX';
 
 export type InputTypeShape = {
   chapterSelect: number;
@@ -83,17 +80,3 @@ export enum RecordingStatus {
   paused = 'paused',
   finished = 'finished'
 }
-
-type SourcecastWorkspaceAttr = {
-  readonly audioUrl: string;
-  readonly codeDeltasToApply: CodeDelta[] | null;
-  readonly currentPlayerTime: number;
-  readonly description: string | null;
-  readonly inputToApply: Input | null;
-  readonly playbackData: PlaybackData;
-  readonly playbackDuration: number;
-  readonly playbackStatus: PlaybackStatus;
-  readonly sourcecastIndex: SourcecastData[] | null;
-  readonly title: string | null;
-};
-export type SourcecastWorkspaceState = SourcecastWorkspaceAttr & WorkspaceState;
