+++ src/features/sourceRecorder/sourcecast/SourcecastReducer.ts
@@ -1,17 +1,16 @@
 import { Reducer } from 'redux';
 
-import { SourceActionType } from '../../utils/actionsHelper';
+import { SourceActionType } from '../../../utils/actionsHelper';
 import {
+  SAVE_SOURCECAST_DATA,
   SET_CODE_DELTAS_TO_APPLY,
   SET_CURRENT_PLAYER_TIME,
   SET_INPUT_TO_APPLY,
   SET_SOURCECAST_DATA,
   SET_SOURCECAST_PLAYBACK_DURATION,
-  SET_SOURCECAST_PLAYBACK_STATUS,
-  UPDATE_SOURCECAST_INDEX
-} from '../sourcecast/SourcecastTypes';
-import { SAVE_SOURCECAST_DATA } from '../sourcereel/SourcereelTypes';
-import { SourcecastWorkspaceState } from './SourcecastTypes';
+  SET_SOURCECAST_PLAYBACK_STATUS
+} from '../SourceRecorderTypes';
+import { SourcecastWorkspaceState, UPDATE_SOURCECAST_INDEX } from './SourcecastTypes';
 
 export const SourcecastReducer: Reducer<SourcecastWorkspaceState> = (
   state: SourcecastWorkspaceState,
