+++ src/features/sourceRecorder/sourcecast/__tests__/SourcecastReducer.ts
@@ -1,16 +1,20 @@
-import { defaultWorkspaceManager } from '../../../commons/application/ApplicationTypes';
-import { ExternalLibraryNames } from '../../../commons/application/types/ExternalTypes';
+import { defaultWorkspaceManager } from '../../../../commons/application/ApplicationTypes';
+import { ExternalLibraryNames } from '../../../../commons/application/types/ExternalTypes';
 import {
+  CodeDelta,
+  Input,
+  PlaybackData,
+  PlaybackStatus,
+  SAVE_SOURCECAST_DATA,
   SET_CODE_DELTAS_TO_APPLY,
   SET_INPUT_TO_APPLY,
   SET_SOURCECAST_DATA,
   SET_SOURCECAST_PLAYBACK_DURATION,
   SET_SOURCECAST_PLAYBACK_STATUS,
-  UPDATE_SOURCECAST_INDEX
-} from '../../sourcecast/SourcecastTypes';
-import { SAVE_SOURCECAST_DATA } from '../../sourcereel/SourcereelTypes';
+  SourcecastData
+} from '../../SourceRecorderTypes';
 import { SourcecastReducer } from '../SourcecastReducer';
-import { CodeDelta, Input, PlaybackData, PlaybackStatus, SourcecastData } from '../SourcecastTypes';
+import { UPDATE_SOURCECAST_INDEX } from '../SourcecastTypes';
 
 function generateAction(type: string, payload: any = {}) {
   return {
