+++ src/pages/__tests__/createStore.test.ts
@@ -1,12 +1,13 @@
 import { Variant } from 'js-slang/dist/types';
 
 import { compressToUTF16 } from 'lz-string';
-import { createStore } from '../createStore';
-import { SavedState } from '../localStorage';
+
+import { defaultState, OverallState } from '../../commons/application/ApplicationTypes';
+import { ExternalLibraryName } from '../../commons/application/types/ExternalTypes';
 import { DEFAULT_SOURCE_CHAPTER } from '../../utils/constants';
 import { history } from '../../utils/history';
-import { OverallState, defaultState } from '../../commons/application/ApplicationTypes';
-import { ExternalLibraryName } from '../../commons/application/types/ExternalTypes';
+import { createStore } from '../createStore';
+import { SavedState } from '../localStorage';
 
 const mockChangedStoredState: SavedState = {
   session: {
