+++ src/utils/actionsHelper.ts
@@ -7,8 +7,9 @@ import * as CollabEditingActions from '../commons/collabEditing/CollabEditingAct
 import * as WorkspaceActions from '../commons/workspace/WorkspaceActions';
 import * as DashboardActions from '../features/dashboard/DashboardActions';
 import * as PlaygroundActions from '../features/playground/PlaygroundActions';
-import * as SourcecastActions from '../features/sourcecast/SourcecastActions';
-import * as SourcereelActions from '../features/sourcereel/SourcereelActions';
+import * as SourcecastActions from '../features/sourceRecorder/sourcecast/SourcecastActions';
+import * as SourceRecorderActions from '../features/sourceRecorder/SourceRecorderActions';
+import * as SourcereelActions from '../features/sourceRecorder/sourcereel/SourcereelActions';
 
 import * as GameActions from '../actions/game';
 import * as MaterialActions from '../actions/material';
@@ -22,6 +23,7 @@ export const actions = {
   ...SessionActions,
   ...SourcecastActions,
   ...SourcereelActions,
+  ...SourceRecorderActions,
   ...WorkspaceActions,
 
   ...MaterialActions,
